Move title scene helper module
